Update CONTRIBUTING.md
Added a link to the code of conduct

I removed words like "we need a superb... we need an excellent" because you do not need to be superb nor excellent people. We need nice people who write nice words and we can work with them to make great posts. It's kind of like when someone tells me I need to be "super passionate" about my job all the time to get hired and I'm like "well sometimes im tired and bored, actually".